Reworked

+++ backend/src/main/java/com/v2gdemo/backend/entity/Transaction.java
@@ -16,7 +16,7 @@ public class Transaction {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double value;
+    private Long value;
 
 
     @ManyToOne(fetch = FetchType.LAZY)
